Move upsert queue temporal logic
