+++ front/temporal/upsert_queue/workflows.ts
@@ -1,6 +1,6 @@
 import { proxyActivities } from "@temporalio/workflow";
 
-import type * as activities from "@app/upsert_queue/temporal/activities";
+import type * as activities from "@app/temporal/upsert_queue/activities";
 
 const { upsertDocumentActivity } = proxyActivities<typeof activities>({
   startToCloseTimeout: "10 minutes",
